font/fixed: fix float64 to uint32 conversion failing on linux/arm
Contribution from ~cznic at:

https://gitlab.com/cznic/knuth/-/commit/ae0b147fc8ba5da8093ad4b136b0f3a51409ff61

+++ font/fixed/int12_20.go
@@ -30,7 +30,7 @@ func ParseInt12_20(s string) (Int12_20, error) {
 	if err != nil {
 		return 0, err
 	}
-	return Int12_20(f * (1 << 20)), nil
+	return Int12_20(int32(f * (1 << 20))), nil
 }
 
 // Float64 converts the 12.20 fixed-point number to a floating point one.
